refactor(window): move adjust_other_split_cursors_for_event to impl Window
The body only walks per-window splits and view-state maps to shift
cursors that share the edited buffer; it never reaches across windows.
Moving it onto `impl Window` removes one of the last `self.windows
.get_mut(&self.active_window)...` chains from split_actions.rs and lets
a future Window-side `apply_event_to_active_buffer` invoke it directly
on `self`.
